CBLT-323:  merge changes made after release/1.0.1 branch was repaired:
  CBLT-323:  remove LICENSE files, they come from the build repo
  CBLT-226:  windows package now a setup.exe

+++ build/package-win.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 # 
-#  called by:    build/scripts/jenkins/mobile/build_sync_gateway.sh
+#  called by:    build/scripts/jenkins/mobile/build_sync_gateway.bat
 #  with params: 
 #                PREFIX          e.g.,  opt/couchbase-sync-gateway
 #                PREFIXD                ./opt/couchbase-sync-gateway
@@ -20,24 +20,22 @@ PRODUCT_KIND    = "sync-gateway"
 
 PREFIX          = ARGV[0] || "/opt/#{PRODUCT}"
 PREFIXD         = ARGV[1] || "./opt/#{PRODUCT}"
-PRODUCT_VERSION = ARGV[2] || "1.0-1234"
+PRODUCT_VERSION = ARGV[2] || "0.0.0-1234"
 REPO_SHA        = ARGV[3] || "master"
-PLATFORM        = ARGV[4] || `uname -s`.chomp + "-" +  `uname -m`.chomp
-ARCH            = ARGV[5] ||                           `uname -m`.chomp
+PLATFORM        = ARGV[4] || 'windows-x64'
+ARCH            = ARGV[5] || 'x64'
 
-platform = PLATFORM.gsub("Darwin", "macosx")
-
-RELEASE         = PRODUCT_VERSION.split('-')[0]    # e.g., 1.0
+RELEASE         = PRODUCT_VERSION.split('-')[0]    # e.g., 1.0.0
 BLDNUM          = PRODUCT_VERSION.split('-')[1]    # e.g., 1234
 
-PKGNAME="#{PRODUCT}_#{RELEASE}-#{BLDNUM}_#{platform}"
+PKGNAME="setup_#{PRODUCT}_#{RELEASE}-#{BLDNUM}_#{ARCH}.exe"
 product_base_cap = PRODUCT_BASE[0..0].upcase + PRODUCT_BASE[1..-1] # Ex: "Couchbase".
 
 print "\nDEBUG:  0: PREFIX          = ", PREFIX
 print "\nDEBUG:  1: PREFIXD         = ", PREFIXD
 print "\nDEBUG:  2: PRODUCT_VERSION = ", PRODUCT_VERSION
 print "\nDEBUG:  3: REPO_SHA        = ", REPO_SHA
-print "\nDEBUG:  4: platform        = ", platform
+print "\nDEBUG:  4: PLATFORM        = ", PLATFORM
 print "\nDEBUG:  5: ARCH            = ", ARCH
 print "\n"
 print "\nDEBUG:  RELEASE  = ", RELEASE
@@ -47,33 +45,49 @@ print "\nDEBUG:  PKGNAME  = ", PKGNAME
 print "\n"
 
 START_DIR  = Dir.getwd()
-STAGE_DIR = "#{START_DIR}/build/winzip/#{PKGNAME}"
-FileUtils.rm_rf   "#{STAGE_DIR}"
-FileUtils.mkdir_p "#{STAGE_DIR}"
-
+STAGE_DIR = "#{START_DIR}/opt/couchbase-sync-gateway/"
+                                                     #  created by build_sync_gateway.bat
 [["#{START_DIR}", "#{STAGE_DIR}"]].each do |src_dst|
     Dir.chdir(src_dst[0]) do
         ["manifest.txt.tmpl", "manifest.xml.tmpl"].each do |x|
             target = "#{src_dst[1]}/#{x.gsub('.tmpl', '')}"
-            sh %{sed -e s,@@VERSION@@,#{BLDNUM},g                    #{x} |
-                 sed -e s,@@PLATFORM@@,#{platform},g                      |
-                 sed -e s,@@RELEASE@@,#{RELEASE},g                        |
-                 sed -e s,@@REPO_SHA@@,#{REPO_SHA},g                      |
-                 sed -e s,@@PREFIX@@,#{PREFIX},g                          |
-                 sed -e s,@@PRODUCT@@,#{PRODUCT},g                        |
-                 sed -e s,@@PRODUCT_BASE@@,#{PRODUCT_BASE},g              |
-                 sed -e s,@@PRODUCT_BASE_CAP@@,#{product_base_cap},g      |
-                 sed -e s,@@PRODUCT_KIND@@,#{PRODUCT_KIND},g > #{target}}
-            sh %{chmod a+x #{target}}
+            text = File.read("#{x}")
+            text.gsub("@@VERSION@@",          "#{RELEASE}-#{BLDNUM}")
+            text.gsub("@@PLATFORM@@",         "#{PLATFORM}"         )
+            text.gsub("@@RELEASE@@",          "#{RELEASE}"          )
+            text.gsub("@@REPO_SHA@@",         "#{REPO_SHA}"         )
+            text.gsub("@@PREFIX@@",           "#{PREFIX}"           )
+            text.gsub("@@PRODUCT@@",          "#{PRODUCT}"          )
+            text.gsub("@@PRODUCT_BASE@@",     "#{PRODUCT_BASE}"     )
+            text.gsub("@@PRODUCT_BASE_CAP@@", "#{product_base_cap}" )
+            text.gsub("@@PRODUCT_KIND@@",     "#{PRODUCT_KIND}"     )
+            File.open("#{target}", "w") { |file| file.puts text }
         end
     end
 end
-FileUtils.mv  "#{STAGE_DIR}/manifest.txt", "#{PREFIXD}/manifest.txt"
-FileUtils.mv  "#{STAGE_DIR}/manifest.xml", "#{PREFIXD}/manifest.xml"
 
-Dir.chdir("#{START_DIR}") do
-    sh %{tar --directory #{File.dirname(PREFIXD)} -czf "#{PKGNAME}.tar.gz"   #{File.basename(PREFIXD)}}
-    FileUtils.cp                                       "#{PKGNAME}.tar.gz", "#{PREFIXD}/#{PKGNAME}.tar.gz"
+#  make installer
+
+INSTALL_PROJ  = "Sync_Gateway.ism"
+INSTALL_SRC   = "#{START_DIR}/windows/InstallShield_2014_Projects"
+INSTALL_OUT   = "#{INSTALL_SRC}/Sync_Gateway/SINGLE_EXE_IMAGE/Release/DiskImages/DISK1"
+
+proj_param    = "#{INSTALL_SRC}/#{INSTALL_PROJ}"
+proj_param    = proj_param.gsub('/', '\\')
+
+path_to_workspace = "#{ENV['WORKSPACE']}"
+path_to_sgw_files = "#{path_to_workspace}\\app-under-test\\sync_gateway\\build\\opt\\couchbase-sync-gateway"
+installer_params  = "-l PATH_TO_JENKINS_WORKSPACE=#{path_to_workspace} -l PATH_TO_SYNC_GATEWAY_FILES=#{path_to_sgw_files}"
+
+print "\nISCmdBld.exe -v -y #{RELEASE} -d ProductVersion=#{RELEASE}.#{BLDNUM} #{installer_params} -p #{proj_param}\n"
+print   `ISCmdBld.exe -v -y #{RELEASE} -d ProductVersion=#{RELEASE}.#{BLDNUM} #{installer_params} -p #{proj_param}`
+print "\n\n"
+
+if  File.exists?("#{INSTALL_OUT}/setup.exe")
+    FileUtils.cp "#{INSTALL_OUT}/setup.exe",  "#{PREFIXD}/#{PKGNAME}"
+    FileUtils.mv "#{STAGE_DIR}/manifest.txt", "#{PREFIXD}/manifest.txt"
+    FileUtils.mv "#{STAGE_DIR}/manifest.xml", "#{PREFIXD}/manifest.xml"
+else
+    print "\nFAIL:  File does not exist: #{INSTALL_OUT}/setup.exe"
+    exit 99
 end
-
-
